Create subscribers controller
Create controller that has a User model as resource. Also, add logic
behavior: if the user has already signed up, with his email and cpf,
don't throw a error.
Also add the 'thanks' action - which verifies if the user has
subscribed; if not, just go home.

TODO: fix the belongs to (it's not working anymore) and remove those
ugly lines from the create method.