Add subscription error message element and small fixes on the subscription form
We render error messages inside the #subscription_error_message element.
Also, show only numbers on the month select; add data-token to the form
(instead of having it on a meta tag) and add the real subscriptions
fields as hidden fields (so we can detach the others and send the form).